+++ src/module_4/homework/BankSystem.java
@@ -2,8 +2,8 @@ package module_4.homework;
 
 public interface BankSystem {
 
-    void withdrawOfUser(User user, int amount);
-    void fundUser(User user, int amount);
-    void transferMoney(User fromUser, User toUser, int amount);
+    void withdrawOfUser(User user, double amount);
+    void fundUser(User user, double amount);
+    void transferMoney(User fromUser, User toUser, double amount);
     void paySalary(User user);
 }
